backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
beautifulsoup4==4.12.2
PyPDF2==3.0.1
numpy==1.26.2
python-dotenv==1.0.0
pydantic>=2.7.4,<3.0.0
aiofiles==23.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pytest==7.4.3
httpx>=0.27,<0.29
selenium==4.15.2
webdriver-manager==4.0.1
ghostscript
pdfplumber==0.10.3
openpyxl==3.1.2
thefuzz==0.22.1
tabula-py==2.9.0
requests>=2.31.0
tqdm>=4.66.3
opencv-python==4.9.0.80
jpype1
fuzzywuzzy
camelot-py==0.11.0
flask==3.0.3
flask-cors==4.0.0
pandas==2.1.3
transformers==4.38.2
--find-links https://download.pytorch.org/whl/cu118
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
sentence-transformers==2.5.1
faiss-cpu==1.7.4
langchain
langchain-ollama
langchain_community
langchain-core
langgraph
